server/requirements.txt
Flask
Flask-MySQLdb
python-decouple
Flask-CORS